archive: Change workflow
To see what has changed in the archive files by a commit, extract the
files of the archive into "archive" directory. The archive tag with
its content is remove in the armhf-sunxi-bananapi-r1.xml file to
ensure changes are not forgotten to be added to the zipped
archive. Henceforth the archive needs to be added to the xml file
before executing "elbe initvm submit". Therefore the following commit
is required:

  elbe chg_archive armhf-sunxi-bananapi-r1.xml archive

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>