archive: Execute socat: Use a systemd service
systemd-udevd kills the udev command after a defined time, because it
tooks to long to execute. Therefore create a udev rule, which is
connected to a systemd service which then starts the mirror_serial
script. Every testbox should only forward a single serial
stream. Therefore use only the forwarding of ttyUSB1 which is required
for the current testbox setup. The management of variants will be
added later.

Furthermore, remove the background execution of socat command in
mirror_serial script.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>